Restructure filws.
