+++ core/java/src/main/java/org/phylospec/typeresolver/properties/FromNexusHook.java
@@ -4,7 +4,6 @@ import java.util.Map;
 import java.util.Set;
 import org.phylospec.components.ParsedType;
 import org.phylospec.typeresolver.ResolvedType;
-import org.phylospec.typeresolver.TypePropertyResolverHook;
 
 public class FromNexusHook implements TypePropertyResolverHook {
