+++ core/java/src/main/resources/META-INF/services/org.phylospec.typeresolver.properties.TypePropertyResolverHook
